utils: Move struct cu_files and debug_entry to its header
It'd be better to move "struct cu_files" and "struct debug_entry" to its
header so that other tools can just include "utils/dwarf.h".
